planner-api: Split the main into commands
This commit splits the main file into commands to have another command
_migrate_ along with the _run_ command.
The _migrate_ command is needed to migrate db during unit tests.
