+++ cmd/planner-api/main.go
@@ -1,85 +1,10 @@
 package main
 
-import (
-	"context"
-	"net"
-	"os"
-	"os/signal"
-	"syscall"
-
-	"github.com/kubev2v/migration-planner/internal/api_server/agentserver"
-	"github.com/kubev2v/migration-planner/internal/config"
-	"github.com/kubev2v/migration-planner/pkg/log"
-	"github.com/sirupsen/logrus"
-
-	apiserver "github.com/kubev2v/migration-planner/internal/api_server"
-
-	"github.com/kubev2v/migration-planner/internal/store"
-)
+import "os"
 
 func main() {
-	log := log.InitLogs()
-	log.Println("Starting API service")
-	defer log.Println("API service stopped")
-
-	cfg, err := config.LoadOrGenerate(config.ConfigFile())
+	err := rootCmd.Execute()
 	if err != nil {
-		log.Fatalf("reading configuration: %v", err)
+		os.Exit(1)
 	}
-	log.Printf("Using config: %s", cfg)
-
-	logLvl, err := logrus.ParseLevel(cfg.Service.LogLevel)
-	if err != nil {
-		logLvl = logrus.InfoLevel
-	}
-	log.SetLevel(logLvl)
-
-	log.Println("Initializing data store")
-	db, err := store.InitDB(cfg, log)
-	if err != nil {
-		log.Fatalf("initializing data store: %v", err)
-	}
-
-	store := store.NewStore(db, log.WithField("pkg", "store"))
-	defer store.Close()
-
-	if err := store.InitialMigration(); err != nil {
-		log.Fatalf("running initial migration: %v", err)
-	}
-
-	ctx, cancel := signal.NotifyContext(context.Background(), os.Interrupt, syscall.SIGHUP, syscall.SIGTERM, syscall.SIGQUIT)
-	go func() {
-		listener, err := newListener(cfg.Service.Address)
-		if err != nil {
-			log.Fatalf("creating listener: %s", err)
-		}
-
-		server := apiserver.New(log, cfg, store, listener)
-		if err := server.Run(ctx); err != nil {
-			log.Fatalf("Error running server: %s", err)
-		}
-		cancel()
-	}()
-
-	go func() {
-		listener, err := newListener(cfg.Service.AgentEndpointAddress)
-		if err != nil {
-			log.Fatalf("creating listener: %s", err)
-		}
-
-		agentserver := agentserver.New(log, cfg, store, listener)
-		if err := agentserver.Run(ctx); err != nil {
-			log.Fatalf("Error running server: %s", err)
-		}
-		cancel()
-	}()
-
-	<-ctx.Done()
-}
-
-func newListener(address string) (net.Listener, error) {
-	if address == "" {
-		address = "localhost:0"
-	}
-	return net.Listen("tcp", address)
 }
